Changes to make the build system vpath friendly.
Remove embedded hwloc and plpa libraries, which we don't really need
to make hydra work and which would require a bunch of work to make 
PAC_CONFIG_SUBDIR_ARGS work in our vpath setup.  Since hydra supports
using an external hwloc, this was the solution with minimal changes
for hwloc and plpa.

mpl is assumed to be able to configure, so we just switch from
PAC_CONFIG_SUBDIRS to AC_CONFIG_SUBDIRS and call it a day.